+++ CHANGELOG.md
@@ -14,6 +14,7 @@ In Development
   20, whichever is lower
 - Add `--ok-errors` option
 - Store start & successful end times of program runs in `.s3invsync.state.json`
+- Support objects without version IDs
 
 v0.1.0-alpha.2 (2025-01-06)
 ---------------------------